# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ProjectOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: Projects list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProject"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProject"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks with optional filters
      parameters:
        - name: projectId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: assigneeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: priority
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Tasks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTask"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTask"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tasks/reorder:
    post:
      operationId: reorderTasks
      tags: [tasks]
      summary: Reorder tasks (for kanban drag-and-drop)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderTasks"
      responses:
        "200":
          description: Reordered
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
  /sprints:
    get:
      operationId: listSprints
      tags: [sprints]
      summary: List sprints
      parameters:
        - name: projectId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Sprints list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sprint"
    post:
      operationId: createSprint
      tags: [sprints]
      summary: Create a sprint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSprint"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sprint"
  /sprints/{id}:
    patch:
      operationId: updateSprint
      tags: [sprints]
      summary: Update a sprint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSprint"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sprint"
  /time-entries:
    get:
      operationId: listTimeEntries
      tags: [timeEntries]
      summary: List time entries
      parameters:
        - name: projectId
          in: query
          schema:
            type: ["integer", "null"]
        - name: memberId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Time entries list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeEntry"
    post:
      operationId: createTimeEntry
      tags: [timeEntries]
      summary: Create a time entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimeEntry"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
  /time-entries/{id}:
    patch:
      operationId: updateTimeEntry
      tags: [timeEntries]
      summary: Update a time entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTimeEntry"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
    delete:
      operationId: deleteTimeEntry
      tags: [timeEntries]
      summary: Delete a time entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        client:
          type: string
        budget:
          type: number
        health:
          type: integer
        phase:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - icon
        - color
        - client
        - budget
        - health
        - phase
        - createdAt
    CreateProject:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        client:
          type: string
        budget:
          type: number
        health:
          type: integer
        phase:
          type: string
      required:
        - name
        - icon
        - color
        - client
        - budget
    UpdateProject:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        client:
          type: string
        budget:
          type: number
        health:
          type: integer
        phase:
          type: string
...
    Subtask:
      type: object
      properties:
        title:
          type: string
        done:
          type: boolean
      required:
        - title
        - done
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [task, bug, feature, story, epic]
        status:
          type: string
          enum: [backlog, todo, inprogress, review, done, blocked]
        priority:
          type: string
          enum: [critical, high, medium, low]
        projectId:
          type: integer
        sprintId:
          type: ["integer", "null"]
        assigneeIds:
          type: array
          items:
            type: integer
        points:
          type: integer
        due:
          type: ["string", "null"]
          format: date-time
        tags:
          type: array
          items:
            type: string
        subtasks:
          type: array
          items:
            $ref: "#/components/schemas/Subtask"
        notes:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - type
        - status
        - priority
        - projectId
        - assigneeIds
        - points
        - tags
        - subtasks
        - notes
        - sortOrder
        - createdAt
    CreateTask:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [task, bug, feature, story, epic]
        status:
          type: string
          enum: [backlog, todo, inprogress, review, done, blocked]
        priority:
          type: string
          enum: [critical, high, medium, low]
        projectId:
          type: integer
        sprintId:
          type: ["integer", "null"]
        assigneeIds:
          type: array
          items:
            type: integer
        points:
          type: integer
        due:
          type: ["string", "null"]
          format: date-time
        tags:
          type: array
          items:
            type: string
        subtasks:
          type: array
          items:
            $ref: "#/components/schemas/Subtask"
        notes:
          type: string
      required:
        - title
        - type
        - status
        - priority
        - projectId
    UpdateTask:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [task, bug, feature, story, epic]
        status:
          type: string
          enum: [backlog, todo, inprogress, review, done, blocked]
        priority:
          type: string
          enum: [critical, high, medium, low]
        projectId:
          type: integer
        sprintId:
          type: ["integer", "null"]
        assigneeIds:
          type: array
          items:
            type: integer
        points:
          type: integer
        due:
          type: ["string", "null"]
          format: date-time
        tags:
          type: array
          items:
            type: string
        subtasks:
          type: array
          items:
            $ref: "#/components/schemas/Subtask"
        notes:
          type: string
        sortOrder:
          type: integer
    ReorderTasks:
      type: object
      properties:
        taskIds:
          type: array
          items:
            type: integer
        status:
          type: string
          enum: [backlog, todo, inprogress, review, done, blocked]
      required:
        - taskIds
        - status
    Sprint:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        projectId:
          type: integer
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        goal:
          type: string
        status:
          type: string
          enum: [planned, active, completed]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - projectId
        - startDate
        - endDate
        - goal
        - status
        - createdAt
    CreateSprint:
      type: object
      properties:
        name:
          type: string
        projectId:
          type: integer
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        goal:
          type: string
        status:
          type: string
          enum: [planned, active, completed]
      required:
        - name
        - projectId
        - startDate
        - endDate
        - goal
    UpdateSprint:
      type: object
      properties:
        name:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        goal:
          type: string
        status:
          type: string
          enum: [planned, active, completed]
    TimeEntry:
      type: object
      properties:
        id:
          type: integer
        memberId:
          type: integer
        projectId:
          type: integer
        description:
          type: string
        hours:
          type: number
        date:
          type: string
          format: date-time
        billable:
          type: boolean
        rate:
          type: number
        amount:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - memberId
        - projectId
        - description
        - hours
        - date
        - billable
        - rate
        - amount
        - createdAt
    CreateTimeEntry:
      type: object
      properties:
        memberId:
          type: integer
        projectId:
          type: integer
        description:
          type: string
        hours:
          type: number
        date:
          type: string
          format: date-time
        billable:
          type: boolean
        rate:
          type: number
      required:
        - memberId
        - projectId
        - description
        - hours
        - date
        - billable
        - rate
    UpdateTimeEntry:
      type: object
      properties:
        description:
          type: string
        hours:
          type: number
        billable:
          type: boolean
...