# lib/api-zod/src/generated/types/updateTask.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ProjectOS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Subtask } from "./subtask";
import type { UpdateTaskPriority } from "./updateTaskPriority";
import type { UpdateTaskStatus } from "./updateTaskStatus";
import type { UpdateTaskType } from "./updateTaskType";

export interface UpdateTask {
  title?: string;
  type?: UpdateTaskType;
  status?: UpdateTaskStatus;
  priority?: UpdateTaskPriority;
  projectId?: number;
  /** @nullable */
  sprintId?: number | null;
  assigneeIds?: number[];
  points?: number;
  /** @nullable */
  due?: Date | null;
  tags?: string[];
  subtasks?: Subtask[];
  notes?: string;
  sortOrder?: number;
}