# lib/api-zod/src/generated/types/updateTimeEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ProjectOS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateTimeEntry {
  description?: string;
  hours?: number;
  billable?: boolean;
}